// src/app/Types/TypeCompanies.ts
export interface TypeCompanies{
    CompSno: number;
    Comp_Code: string;
    Comp_Name: string;
    Fin_From: number;
    Fin_To: number;
    Books_From: number;
    Address1: string;
    Address2: string;
    Address3: string;
    City: string;
    State: string;
    Pincode: string;
    Email: string;
    Phone: string;
    License_No: string;
    Hide_Status: number;
    App_Version: number;
    Db_Version: number;
    Status: number;
    CommMasters: number;
}